Improve order management
